feat(ical): rebuild the multi-feed UI on the dashboard's own primitives
The iCal routes hand-rolled their own layout while the rest of the dashboard
composes NavigationMenu cards, RouteShell, editable items, MetadataRow and
Pagination. The feed detail page now follows the calendar detail page: a title
with its calendar count, sections of menu rows, an information block and a
destructive row at the end. Copy that narrated a link's history is gone, and the
legacy alias URL is no longer surfaced at all.

Also restores the per-feed event name template dropped in the multi-feed port,
and makes the list's calendar count an integer instead of a bigint string, which
pluralized as "1 calendars".

+++ services/api/src/utils/ical-feeds.ts
@@ -198,7 +198,7 @@ const buildFeedListQuery = (client: Pick<DatabaseClient, "select">, userId: stri
   client
     .select({
       ...FEED_ROW_COLUMNS,
-      calendarCount: sql<number>`count(${icalFeedCalendarsTable.calendarId})`,
+      calendarCount: sql<number>`count(${icalFeedCalendarsTable.calendarId})::int`,
     })
     .from(icalFeedsTable)
     .leftJoin(icalFeedCalendarsTable, eq(icalFeedCalendarsTable.feedId, icalFeedsTable.id))
